move code around to be clearer
