+++ src/actions/LoadFileDialog.tsx
@@ -1,10 +1,10 @@
-import './dialog.css'
+import '../dialog.css'
 
 import {DialogContext, Spacer} from "josh_react_util"
 import React, {ChangeEvent, useContext, useRef, useState} from "react"
 
-import {loadPNGJSON} from "./exporters/json"
-import {GlobalState} from "./models/state"
+import {loadPNGJSON} from "../exporters/json"
+import {GlobalState} from "../models/state"
 
 function isValidJSONPNGFile(file: File) {
     if(!file) return false
