+++ src/actions/reactactions.tsx
@@ -2,14 +2,14 @@ import {Size} from "josh_js_util"
 import {DialogContext, Spacer} from "josh_react_util"
 import React, {ChangeEvent, JSX, useContext, useState} from "react"
 
-import {IconButton, ReactMenuAction} from "./common"
-import {loadPNGJSON} from "./exporters/json"
-import {SupportedIcons} from "./icons"
+import {IconButton, ReactMenuAction} from "../common"
+import {loadPNGJSON} from "../exporters/json"
+import {SupportedIcons} from "../icons"
+import {DocClass, DocDef, PageDef} from "../models/om"
+import {GlobalState} from "../models/state"
+import {lookup_name, Unit} from "../models/unit"
 import {ListFilesDialog} from "./ListFilesDialog"
 import {LoadFileDialog} from "./LoadFileDialog"
-import {DocClass, DocDef, PageDef} from "./models/om"
-import {GlobalState} from "./models/state"
-import {lookup_name, Unit} from "./models/unit"
 import {SettingsDialog} from "./SettingsDialog"
 
 function UploadButton(props: { state: GlobalState }) {
