+++ src/propsheet/ProxySelectionDialog.css
